resolve resolution problem
