+++ src/App.tsx
@@ -1,13 +1,13 @@
 import { Provider } from "react-redux";
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
 
-import MyCO2 from "./components/MyCO2";
 import Co2ImageTrack from "./components/app/co2imagetrack";
 import Co2RealtimeTrack from "./components/app/co2realtimetrack";
 import Layout from "./components/app/layout";
 import NotFound from "./pages/404";
 import Home from "./pages/landing";
 import { store } from "./rtk/store";
+import MyCo2Track from "./components/app/myco2track";
 
 const router = createBrowserRouter([
   {
@@ -27,11 +27,11 @@ const router = createBrowserRouter([
         path: "image",
         element: <Co2ImageTrack />,
       },
+      {
+        path: "myco2",
+        element: <MyCo2Track />,
+      },
     ],
-  },
-  {
-    path: "myco2",
-    element: <MyCO2 />,
   },
 ]);
