+++ src/components/app/layout/index.tsx
@@ -12,10 +12,18 @@ function Layout() {
           <Center fontSize="3xl" fontWeight="bold" mt="10" color="#2BD575">
             CO2Track
           </Center>
-          <Center fontSize="xl" mt="5">
+          <Center fontSize="lg" mt="5">
             Vous permet de visualiser l'analyse d'une vue d'ensemble en temps
             réel pour reconnaitre le total d'une Emission de CO2 dégager dans
             une salle.
+          </Center>
+          <Center fontSize="3xl" fontWeight="bold" mt="10" color="#2BD575">
+            MyCO2
+          </Center>
+          <Center fontSize="lg" mt="5">
+            Vous offre une vue d'ensemble de vos émissions de carbone. Tout est
+            regroupé pour vous aider à visualiser les moyens existant pour
+            connaitre les résultats de myCO2.
           </Center>
         </Stack>
       </Box>
