+++ src/components/app/navbar/index.tsx
@@ -14,7 +14,7 @@ import User from "../../../assets/user.png";
 const links = [
   { name: "En temps réel", href: "/co2track/realtime" },
   { name: "Par image", href: "/co2track/image" },
-  { name: "MyCO2", href: "#myco2" },
+  { name: "MyCO2", href: "/co2track/myco2" },
 ];
 
 const Navbar = () => {
